tests: Create test_potential fixture
